Split LocalRoomName and LocalRoomPosition into separate files
