+++ src/local.rs
@@ -1,0 +1,5 @@
+//! Pure-data structures relating to Screeps.
+mod room_name;
+mod room_position;
+
+pub use self::{room_name::*, room_position::*};
